requirements.txt
galileo-db==0.10.5.dev2
edgerun-galileo-faas>=0.0.1.dev8
edgerun-faas>=0.0.1.dev3
telemc>=0.3.0
gunicorn==19.9.0
falcon==2.0.0
requests>=2.28.2
redis>=3.5.3
mysql-connector-python==8.0.16
pymq>=0.4.0
pyyaml>=5.4.1
click>=7.0
influxdb_client>=1.30.0
